requirements.txt
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
python_dateutil==2.8.2
PyYAML==6.0
scikit_learn==0.24.2

